requirements.txt
# API Clients
anthropic>=0.18.0
openai>=1.12.0

# Machine Learning & NLP
sentence-transformers==3.3.1
torch>=2.6.0
numpy>=1.26.0
scipy>=1.11.0

# Data Processing
pandas>=2.2.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# CLI Framework
click==8.1.7

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
flake8==7.0.0
black==23.12.1
pylint==3.0.3

# Utilities
tqdm==4.66.1
requests==2.31.0

# Additional dependencies
scikit-learn==1.3.2
